starting to fixed excited state code
