+++ src/apps/molresponse/CMakeLists.txt
@@ -1,4 +1,4 @@
 # src/apps/molresponse
 
-add_executable(molresponse molresponse.cc density.cc property.cc TDDFT.cc iterate_freq.cc iterate_xy.cc global_functions.cc basic_operators.cc Plot_VTK.cc polynomial.cc timer.cc gaussian.cc NWChem.cc)
+add_executable(molresponse molresponse.cc density.cc property.cc TDDFT.cc iterate_freq.cc iterate_xy.cc iterate_excited.cc global_functions.cc basic_operators.cc Plot_VTK.cc polynomial.cc timer.cc gaussian.cc NWChem.cc)
 target_link_libraries(molresponse -lstdc++fs MADchem  )
